requirements.txt
streamlit 
yfinance 
matplotlib 
openai
python-dotenv 
requests
crewai
langchain
langchain-community